Consolidated classes for Row and Grid elements
